requirements_.txt
fastapi>=0.68.0
pydantic>=1.8.0
uvicorn>=0.15.0
psycopg2>=2.9
peewee>=2.15
python-dotenv>=1.0
pydantic_settings>=2.0.0